+++ conf.py
@@ -18,11 +18,11 @@ sys.path.insert(0, os.path.abspath('./'))
 # -- Project information -----------------------------------------------------
 
 project = 'nimbusinator'
-copyright = '2019, Tim Adams'
+copyright = '2020, Tim Adams'
 author = 'Tim Adams'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1.0'
+release = '0.2.2'
 
 
 # -- General configuration ---------------------------------------------------